👌 check spec import everywhere + simplification
